ISSUE #4275 - fix collection select
